requirements.txt
certifi==2017.7.27.1
future==0.16.0
pendulum==1.3.0
pkg-resources==0.0.0
python-dateutil==2.6.1
python-telegram-bot==8.0
pytz==2017.2
pytzdata==2017.2.2
six==1.11.0
tzlocal==1.4
virtualenv==15.1.0
